+++ src/main/kotlin/de/itemis/mps/gradle/BackendConfigurations.kt
@@ -4,4 +4,5 @@ internal object BackendConfigurations {
     internal const val GENERATE_BACKEND_CONFIGURATION_NAME = "generateBackend"
     internal const val MODELCHECK_BACKEND_CONFIGURATION_NAME = "modelcheckBackend"
     internal const val EXECUTE_BACKEND_CONFIGURATION_NAME = "executeBackend"
+    internal const val REMIGRATE_BACKEND_CONFIGURATION_NAME = "remigrateBackend"
 }