+++ CHANGELOG.md
@@ -2,6 +2,16 @@
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## 1.25.0
+
+### Added
+
+- `Remigrate` task type to run re-runnable migrations on a project or several projects.
+
+### Changed
+
+- Extensions now use `--plugin-root` command line argument of backends, requiring backend versions 1.15.0 or above.
+
 ## 1.24.0
 
 ### Changed